+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 7.0 [not yet released]
 
 - remove StringEncodedValue support from custom model due to insufficient usage/testing
+- handle also node_tags in handleWayTags, when extending AbstractAccessParser call handleNodeTags, #2738
 - Format of 'areas' in CustomModel changed to 'FeatureCollection'. The old format is deprecated and will be removed in a later version, #2734
 - TagParser#handleWayTags no longer returns an IntsRef. We assume it never returned anything other than the input IntsRef.
 - there is no longer a default value for the distanceInfluence parameter in custom models sent via client-hc. Previously it was 70. Not setting it explicitly now means the server-side value will be used. getDistanceInfluence can now be null. Server-side profiles with custom weighting now use distance_influence: 0 by default (previously it was 70). see #2716
@@ -9,11 +10,12 @@
 - the parameters vehicle, weighting, edge_based and turn_costs are no longer supported, use the profile parameter instead
 - removed motorroad to road_class conversion, #2329
 - removed YAML support for custom models on the server-side. Only allow JSON with // comments.
-- Bike2WeightTagParser was removed. Use the bike vehicle with a custom model, see custom_models/bike2.json
+- Bike2WeightTagParser was removed. Use the bike vehicle with a custom model, see custom_models/bike.json
 - CurvatureWeighting was removed. Use a custom model with 'curvature' instead, see custom_models/curvature.json (#2665)
 - internal keys for EdgeKVStorage changed to contain the street_ prefix like the path details too. Similarly, the
   extra_info in the instructions of the API response, see #2661
 - subnetwork preparation can now be run in parallel to slightly speed up the base graph import (#2737)
+- The block_area parameter was removed. Use custom model areas instead.
 
 ### 6.0 [13 Sep 2022]